2005-01-22	Ralf Corsepius <ralf.corsepius@rtems.org>
	* tod/todcfg.c: size_t RTC_Count.

+++ c/src/lib/libbsp/powerpc/shared/tod/todcfg.c
@@ -44,7 +44,7 @@ rtc_tbl RTC_Table[] = {
 
 #define NUM_RTCS (sizeof(RTC_Table)/sizeof(rtc_tbl))
 
-unsigned long RTC_Count = NUM_RTCS;
+size_t RTC_Count = NUM_RTCS;
 
 rtems_device_minor_number RTC_Minor;
